+++ src/core/managers/index.ts
@@ -1,5 +1,1 @@
-export * from './LedgerConnectStateManager';
-export * from './SignTransactionsStateManager';
 export * from './TransactionManager';
-export * from './WalletConnectStateManager';
-export * from './PendingTransactionsStateManager';